requirements.txt
Flask==2.0.2
Werkzeug==2.0.2
boto3==1.18.65
Pillow==8.4.0
python-decouple
